Feature : 질문 답변 관리자 삭제 기능 (#291)
* feat: 관리자 질문 삭제 권한 추가

* feat: 관리자 답변 삭제 기능 추가

* update submodule

* test: 깨지는 테스트 수정

* refactor: 관련없는 파일 rollback 및 system.out 삭제

* refactor: 인터셉터 Role 정책 수정

* refactor: validator return 값 변경

* refactor: delete 메서드에 role 표기

* refactor: footer height 조정

+++ src/main/java/underdogs/devbie/answer/service/AnswerService.java
@@ -80,9 +80,15 @@ public class AnswerService {
     public void delete(User user, Long id) {
         Answer answer = readOne(id);
 
-        validateAuthentication(user, answer);
+        validateAuthenticationOrAdmin(user, answer);
 
         answerRepository.deleteById(answer.getId());
+    }
+
+    private void validateAuthenticationOrAdmin(User user, Answer answer) {
+        if (isNotAuthorOfQuestion(user, answer) && user.isNotAdmin()) {
+            throw new NotMatchedAnswerAuthorException();
+        }
     }
 
     public AnswerResponses readByQuestionId(Long questionId) {
